refactor(api): extract AppModelConfig TypedDict definitions into separate module
